app6/requirements.txt
flask
pymilvus
sentence-transformers
llama-cpp-python
langchain-mcp
langchain-mcp-adapters
langchain
aiohttp